+++ index.js
@@ -9,25 +9,14 @@ module.exports = {
     "stylelint-config-recommended-scss",
   ],
   "rules": {
-    "indentation": 4, // deprecated
-    "max-empty-lines": 2, // deprecated
-    "no-eol-whitespace": true, // deprecated
-    "no-missing-end-of-source-newline": true, // deprecated
-    "declaration-colon-newline-after": "always-multi-line", // deprecated
     "declaration-empty-line-before": "never",
-    "value-list-max-empty-lines": 0, // deprecated
 
-    "selector-pseudo-element-colon-notation": "double", // deprecated
     "color-named": "never",
 
     "declaration-no-important": true,
     "declaration-block-no-duplicate-properties": true,
 
-    "property-case": "lower", // deprecated
     "value-keyword-case": "lower",
-    "unit-case": "lower", // deprecated
-    "color-hex-case": "lower", // deprecated
-    "selector-pseudo-element-case": "lower", // deprecated
 
     "selector-no-vendor-prefix": true,
     "property-no-vendor-prefix": true,
@@ -35,8 +24,6 @@ module.exports = {
     "media-feature-name-no-vendor-prefix": true,
     "at-rule-no-vendor-prefix": true,
 
-    "number-leading-zero": "never", // deprecated
-    "number-no-trailing-zeros": true, // deprecated
     "function-calc-no-unspaced-operator": true,
     "scss/operator-no-newline-before": true,
     "scss/operator-no-newline-after": true,
@@ -58,12 +45,6 @@ module.exports = {
         "extend",
       ],
     }],
-    // deprecated
-    "block-closing-brace-newline-after": [
-      "always", {
-        "ignoreAtRules": ["if", "else"]
-      }
-    ],
     "scss/at-else-closing-brace-newline-after": "always-last-in-chain",
     "scss/at-else-closing-brace-space-after": "always-intermediate",
     "scss/at-else-empty-line-before": "never",